tests/requirements.txt
parameterized==0.8.1
pytest-order==1.0.1
urllib3==1.26.9
pytest-html==3.1.1
pytest-html-reporter==0.2.9
pytest-parametrized==1.3
python-dateutil==2.8.2
pytest==7.0.1
requests-toolbelt==0.10.0
#Pillow==9.3
xlrd==1.2.0
hurry.filesize==0.9
pytest-excel==1.5.0
pyproject-toml==0.0.10